feat: serde support for `JsonPathQuery` and `MainEngine`
Implemented `serde::Serialize` and `serde::Deserialize`
for `MainEngine` in rsonpath-lib, and `JsonPathQuery`
in rsonpath-syntax with all of its constituent substructs.
The `serde` dependency is guarded behind an optional feature.

To properly proptest this feature `rsonpath-lib` needs access
to the arbitrary query generation from `rsonpath-syntax`.
The cleanest way to do this is to extract the logic to a separate
crate, giving rise to `rsonpath-syntax-proptest`.

The serialization format is not stable, as the `Automaton`
is expected to evolve. Thus, serialization includes a version
and deserialization will fail if the version disagrees.

Ref: #624
