+++ crates/rsonpath-lib/src/string_pattern.rs
@@ -3,6 +3,7 @@ use rsonpath_syntax::str::JsonString;
 /// String pattern coming from a JSONPath query that can be matched against strings in a JSON.
 ///
 /// Right now the only pattern is matching against a given [`JsonString`].
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Debug, Clone)]
 pub struct StringPattern(JsonString);
