+++ crates/rsonpath-test/Cargo.toml
@@ -14,8 +14,8 @@ rust-version = "1.70.0"
 publish = false
 
 [dependencies]
-serde = { version = "1.0.216", features = ["derive"] }
-serde_json = "1.0.133"
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
