+++ fuzz/Cargo.toml
@@ -10,12 +10,11 @@ cargo-fuzz = true
 [dependencies]
 arbitrary = { version = "1.4.*", features = ["derive"] }
 libfuzzer-sys = { version = "0.4.8" }
-serde = "1.0.216"
+serde = "1.0.217"
 serde_json = "1.0.134"
 
 [dependencies.rsonpath-lib]
 path = "../crates/rsonpath-lib"
-features = ["arbitrary"]
 
 [dependencies.rsonpath-syntax]
 path = "../crates/rsonpath-syntax"
